requirements.txt
flask
flask-sqlalchemy
pytest
pytest-flask
pytest-cov
pytest-flask-sqlalchemy
black